requirements.txt
arvados-cwl-runner==2.0.3.1
arvados-python-client==2.0.2
bagit==1.7.0
biopython==1.76
CacheControl==0.11.7
certifi==2020.4.5.1
chardet==3.0.4
ciso8601==2.1.3
click==7.1.2
coloredlogs==14.0
cwl-runner==1.0
cwltool==1.0.20190831161204
decorator==4.4.2
future==0.18.2
google-api-python-client==1.6.7
httplib2==0.17.3
humanfriendly==8.2
idna==2.9
isodate==0.6.0
lockfile==0.12.2
lxml==4.5.1
mistune==0.8.4
mypy-extensions==0.4.3
networkx==2.2
numpy==1.18.4
oauth2client==4.1.3
pathlib2==2.3.5
prov==1.5.1
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.23.0
rsa==4.0
ruamel.yaml==0.15.77
scandir==1.10.0
schema-salad==4.5.20190815125611
shellescape==3.4.1
six==1.14.0
typing==3.7.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
ws4py==0.5.1
